pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "user-service"
version = "0.1.0"
description = "Standalone reusable FastAPI user service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.16.0",
  "pydantic-settings>=2.10.0",
  "passlib[bcrypt]>=1.7.4",
  "pillow>=11.3.0",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.20",
]

[project.optional-dependencies]
test = [
  "pytest>=8.4.0",
  "pytest-cov>=6.2.0",
  "httpx>=0.28.0",
]

[tool.uv]
dev-dependencies = []

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
